# [optimization]: make Var.t.value unboxed and specialize cutoff to avoid indirect calls in the majority case(which is physical equality)