I've added the benchmarks below and they show decent improvement over the current version(for which the benchmarks lie [here](https://github.com/ocaml-multicore/par_incr/pull/2) ). 


## filter

```bash
NUM_DOMAINS=2 RUNS=10 bash bench_runner.sh filter.exe 1000 100000 "i*10" 1
+ dune exec --release -- test/filter.exe -r 10 -n 1000 -c 10 -d 2
# Filtering list of size: 1000 | 10 elements changed in propagation | Domains spawned: 2
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 0.01261 ms         0.01144 ms         10       0.02598 ms         0.007152 ms
incr-seq-filter-initial-cons  0.2218 ms          0.2089 ms          10       0.3130 ms          0.1950 ms
incr-par-filter-initial-cons  0.3302 ms          0.3160 ms          10       0.4730 ms          0.2901 ms
current-incr-filter-init-cons 0.4548 ms          0.4354 ms          10       0.5970 ms          0.3762 ms
js-incr-filter-init-cons      0.5043 ms          0.4869 ms          10       0.6849 ms          0.4189 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
js-incr-filter-change-prop      0.01709 ms         0.01716 ms         10       0.02479 ms         0.006198 ms
incr-seq-filter-change-prop     0.01826 ms         0.01788 ms         10       0.02980 ms         0.008106 ms
incr-par-filter-change-prop     0.02560 ms         0.02503 ms         10       0.04005 ms         0.01692 ms
current-incr-filter-change-prop 0.04355 ms         0.04434 ms         10       0.05602 ms         0.02193 ms
+ dune exec --release -- test/filter.exe -r 10 -n 10000 -c 100 -d 2
# Filtering list of size: 10000 | 100 elements changed in propagation | Domains spawned: 2
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 0.1069 ms          0.09691 ms         10       0.1490 ms          0.09489 ms
incr-par-filter-initial-cons  9.7951 ms          9.7054 ms          10       10.9479 ms         9.1121 ms
incr-seq-filter-initial-cons  9.7994 ms          10.5044 ms         10       11.6369 ms         7.2729 ms
js-incr-filter-init-cons      13.5192 ms         13.2409 ms         10       15.9008 ms         12.6709 ms
current-incr-filter-init-cons 15.6859 ms         15.9019 ms         10       18.5000 ms         10.3690 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
js-incr-filter-change-prop      0.1663 ms          0.1718 ms          10       0.1988 ms          0.06818 ms
incr-par-filter-change-prop     0.1750 ms          0.1754 ms          10       0.2031 ms          0.1389 ms
incr-seq-filter-change-prop     0.2218 ms          0.2214 ms          10       0.2620 ms          0.1840 ms
current-incr-filter-change-prop 0.6251 ms          0.6095 ms          10       0.8878 ms          0.4270 ms
+ dune exec --release -- test/filter.exe -r 10 -n 100000 -c 1000 -d 2
# Filtering list of size: 100000 | 1000 elements changed in propagation | Domains spawned: 2
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 2.7430 ms          2.5501 ms          10       4.1971 ms          2.4631 ms
incr-par-filter-initial-cons  85.2566 ms         82.9846 ms         10       99.4248 ms         79.7479 ms
incr-seq-filter-initial-cons  86.06116 ms        79.1894 ms         10       105.6149 ms        78.1559 ms
js-incr-filter-init-cons      149.9191 ms        147.2380 ms        10       172.005176 ms      144.6549 ms
current-incr-filter-init-cons 174.2549 ms        172.1415 ms        10       192.6038 ms        156.9950 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
incr-par-filter-change-prop     1.6911 ms          1.6820 ms          10       1.9271 ms          1.4901 ms
js-incr-filter-change-prop      2.3152 ms          2.3045 ms          10       2.8789 ms          1.9359 ms
incr-seq-filter-change-prop     2.9965 ms          3.02851 ms         10       3.2231 ms          2.5010 ms
current-incr-filter-change-prop 13.5433 ms         14.1394 ms         10       15.1331 ms         8.2099 ms


NUM_DOMAINS=4 RUNS=10 bash bench_runner.sh filter.exe 1000 100000 "i*10" 1
+ dune exec --release -- test/filter.exe -r 10 -n 1000 -c 10 -d 4
# Filtering list of size: 1000 | 10 elements changed in propagation | Domains spawned: 4
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 0.01022 ms         0.009059 ms        10       0.01692 ms         0.006914 ms
incr-seq-filter-initial-cons  0.2655 ms          0.2269 ms          10       0.5049 ms          0.2009 ms
incr-par-filter-initial-cons  0.3624 ms          0.3424 ms          10       0.5071 ms          0.2889 ms
current-incr-filter-init-cons 0.4709 ms          0.4464 ms          10       0.6690 ms          0.3478 ms
js-incr-filter-init-cons      0.5165 ms          0.4999 ms          10       0.6370 ms          0.4150 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
incr-seq-filter-change-prop     0.01704 ms         0.01502 ms         10       0.02789 ms         0.01001 ms
js-incr-filter-change-prop      0.01754 ms         0.01692 ms         10       0.02384 ms         0.006914 ms
incr-par-filter-change-prop     0.02851 ms         0.02658 ms         10       0.04696 ms         0.02002 ms
current-incr-filter-change-prop 0.04830 ms         0.05090 ms         10       0.06294 ms         0.02002 ms
+ dune exec --release -- test/filter.exe -r 10 -n 10000 -c 100 -d 4
# Filtering list of size: 10000 | 100 elements changed in propagation | Domains spawned: 4
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 0.2769 ms          0.1069 ms          10       1.8198 ms          0.09679 ms
incr-par-filter-initial-cons  7.5495 ms          7.4065 ms          10       9.7811 ms          6.3300 ms
incr-seq-filter-initial-cons  9.1079 ms          8.7854 ms          10       12.2442 ms         6.9589 ms
current-incr-filter-init-cons 13.4696 ms         13.7075 ms         10       16.08204 ms        9.1338 ms
js-incr-filter-init-cons      14.6072 ms         14.3885 ms         10       16.3509 ms         13.7469 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
incr-par-filter-change-prop     0.1436 ms          0.1399 ms          10       0.1649 ms          0.1130 ms
js-incr-filter-change-prop      0.1730 ms          0.1820 ms          10       0.1931 ms          0.07820 ms
incr-seq-filter-change-prop     0.2136 ms          0.2160 ms          10       0.2431 ms          0.1590 ms
current-incr-filter-change-prop 0.6144 ms          0.6515 ms          10       0.6701 ms          0.4081 ms
+ dune exec --release -- test/filter.exe -r 10 -n 100000 -c 1000 -d 4
# Filtering list of size: 100000 | 1000 elements changed in propagation | Domains spawned: 4
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 2.6028 ms          2.5619 ms          10       3.09705 ms         2.3748 ms
incr-par-filter-initial-cons  75.4290 ms         76.7600 ms         10       83.5659 ms         64.9430 ms
incr-seq-filter-initial-cons  86.5662 ms         89.9484 ms         10       100.7888 ms        70.8792 ms
js-incr-filter-init-cons      154.6686 ms        151.5344 ms        10       174.07202 ms       146.4190 ms
current-incr-filter-init-cons 159.8216 ms        151.1025 ms        10       222.03707 ms       139.3990 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
incr-par-filter-change-prop     1.4409 ms          1.2865 ms          10       2.8970 ms          1.1761 ms
js-incr-filter-change-prop      2.3307 ms          2.2724 ms          10       2.7718 ms          2.1271 ms
incr-seq-filter-change-prop     2.9726 ms          2.9824 ms          10       3.2739 ms          2.5830 ms
current-incr-filter-change-prop 12.8150 ms         13.3589 ms         10       14.5158 ms         8.2941 ms


NUM_DOMAINS=6 RUNS=10 bash bench_runner.sh filter.exe 1000 100000 "i*10" 1
+ dune exec --release -- test/filter.exe -r 10 -n 1000 -c 10 -d 6
# Filtering list of size: 1000 | 10 elements changed in propagation | Domains spawned: 6
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 0.01111 ms         0.01001 ms         10       0.01692 ms         0.009059 ms
incr-seq-filter-initial-cons  0.2236 ms          0.2214 ms          10       0.2520 ms          0.2100 ms
incr-par-filter-initial-cons  0.3387 ms          0.3325 ms          10       0.3969 ms          0.2949 ms
current-incr-filter-init-cons 0.4947 ms          0.4780 ms          10       0.6229 ms          0.3850 ms
js-incr-filter-init-cons      0.5513 ms          0.5221 ms          10       0.8361 ms          0.4398 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
incr-seq-filter-change-prop     0.01685 ms         0.01502 ms         10       0.02479 ms         0.01192 ms
js-incr-filter-change-prop      0.01809 ms         0.01907 ms         10       0.02503 ms         0.005006 ms
current-incr-filter-change-prop 0.04377 ms         0.04553 ms         10       0.05483 ms         0.02408 ms
incr-par-filter-change-prop     0.1005 ms          0.03695 ms         10       0.6840 ms          0.01788 ms
+ dune exec --release -- test/filter.exe -r 10 -n 10000 -c 100 -d 6
# Filtering list of size: 10000 | 100 elements changed in propagation | Domains spawned: 6
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 0.09877 ms         0.09799 ms         10       0.1029 ms          0.09512 ms
incr-par-filter-initial-cons  7.09166 ms         7.04395 ms         10       9.3801 ms          5.9058 ms
incr-seq-filter-initial-cons  8.3873 ms          7.9244 ms          10       10.7781 ms         6.8418 ms
current-incr-filter-init-cons 12.2898 ms         12.4255 ms         10       14.9829 ms         9.001970 ms
js-incr-filter-init-cons      14.9184 ms         15.1045 ms         10       16.8259 ms         13.5200 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
incr-par-filter-change-prop     0.1406 ms          0.1411 ms          10       0.1721 ms          0.09298 ms
js-incr-filter-change-prop      0.1662 ms          0.1739 ms          10       0.2000 ms          0.07510 ms
incr-seq-filter-change-prop     0.2122 ms          0.2154 ms          10       0.2310 ms          0.1640 ms
current-incr-filter-change-prop 0.6264 ms          0.6549 ms          10       0.7119 ms          0.4680 ms
+ dune exec --release -- test/filter.exe -r 10 -n 100000 -c 1000 -d 6
# Filtering list of size: 100000 | 1000 elements changed in propagation | Domains spawned: 6
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 2.6517 ms          2.6329 ms          10       2.8018 ms          2.5601 ms
incr-par-filter-initial-cons  67.9008 ms         67.5469 ms         10       73.1818 ms         64.6178 ms
incr-seq-filter-initial-cons  81.2323 ms         81.9350 ms         10       91.09115 ms        69.04792 ms
current-incr-filter-init-cons 147.6945 ms        147.4289 ms        10       169.8648 ms        108.9251 ms
js-incr-filter-init-cons      160.08098 ms       156.7355 ms        10       188.6699 ms        152.007102 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
incr-par-filter-change-prop     1.1547 ms          1.1729 ms          10       1.2269 ms          0.9989 ms
js-incr-filter-change-prop      2.2650 ms          2.2519 ms          10       2.7179 ms          1.9118 ms
incr-seq-filter-change-prop     3.03132 ms         3.04043 ms         10       3.2179 ms          2.5951 ms
current-incr-filter-change-prop 13.1239 ms         13.6198 ms         10       14.7159 ms         8.06999 ms


NUM_DOMAINS=8 RUNS=10 bash bench_runner.sh filter.exe 1000 100000 "i*10" 1
+ dune exec --release -- test/filter.exe -r 10 -n 1000 -c 10 -d 8
# Filtering list of size: 1000 | 10 elements changed in propagation | Domains spawned: 8
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 0.01502 ms         0.01204 ms         10       0.03004 ms         0.01001 ms
incr-seq-filter-initial-cons  0.3194 ms          0.2580 ms          10       0.5738 ms          0.2448 ms
current-incr-filter-init-cons 0.4921 ms          0.4868 ms          10       0.6148 ms          0.4107 ms
js-incr-filter-init-cons      0.5582 ms          0.5506 ms          10       0.7138 ms          0.4069 ms
incr-par-filter-initial-cons  0.6927 ms          0.3679 ms          10       2.2411 ms          0.3228 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
js-incr-filter-change-prop      0.02105 ms         0.02133 ms         10       0.03290 ms         0.006914 ms
incr-seq-filter-change-prop     0.02355 ms         0.02408 ms         10       0.03314 ms         0.009059 ms
current-incr-filter-change-prop 0.05378 ms         0.05686 ms         10       0.06508 ms         0.01907 ms
incr-par-filter-change-prop     0.1716 ms          0.04839 ms         10       1.3060 ms          0.01692 ms
+ dune exec --release -- test/filter.exe -r 10 -n 10000 -c 100 -d 8
# Filtering list of size: 10000 | 100 elements changed in propagation | Domains spawned: 8
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 0.1064 ms          0.1065 ms          10       0.1208 ms          0.09703 ms
incr-par-filter-initial-cons  11.2697 ms         9.3514 ms          10       21.4941 ms         6.9038 ms
incr-seq-filter-initial-cons  13.7794 ms         10.5119 ms         10       23.5359 ms         8.4419 ms
current-incr-filter-init-cons 14.8310 ms         14.1965 ms         10       20.6580 ms         9.8860 ms
js-incr-filter-init-cons      15.6469 ms         15.2019 ms         10       18.2869 ms         14.5831 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
js-incr-filter-change-prop      0.1763 ms          0.1875 ms          10       0.2050 ms          0.06318 ms
incr-seq-filter-change-prop     0.2263 ms          0.2365 ms          10       0.2450 ms          0.1459 ms
incr-par-filter-change-prop     0.5275 ms          0.1640 ms          10       2.3131 ms          0.1001 ms
current-incr-filter-change-prop 0.6578 ms          0.6555 ms          10       0.8940 ms          0.3740 ms
+ dune exec --release -- test/filter.exe -r 10 -n 100000 -c 1000 -d 8
# Filtering list of size: 100000 | 1000 elements changed in propagation | Domains spawned: 8
## Initial computation
Name                          Avg                Median             Runs     Max time           Min time
static-filter                 3.7908 ms          3.1166 ms          10       8.5790 ms          2.7520 ms
incr-seq-filter-initial-cons  104.1709 ms        97.5259 ms         10       154.7560 ms        91.7260 ms
js-incr-filter-init-cons      178.3752 ms        175.02701 ms       10       198.9870 ms        164.5741 ms
current-incr-filter-init-cons 183.3004 ms        172.6316 ms        10       279.7880 ms        156.9828 ms
incr-par-filter-initial-cons  189.3139 ms        185.7835 ms        10       291.2251 ms        124.7260 ms
## Change propagation
Name                            Avg                Median             Runs     Max time           Min time
incr-par-filter-change-prop     1.4256 ms          1.1811 ms          10       3.07679 ms         1.008987 ms
js-incr-filter-change-prop      2.5243 ms          2.5200 ms          10       2.9640 ms          1.9319 ms
incr-seq-filter-change-prop     3.06961 ms         3.06904 ms         10       3.6449 ms          2.5959 ms
current-incr-filter-change-prop 35.7146 ms         36.5095 ms         10       54.8360 ms         8.8250 ms


```

## merge_sort

```bash
NUM_DOMAINS=2 RUNS=10 bash bench_runner.sh merge_sort.exe 1000 100000 "i*10" 1
+ dune exec --release -- test/merge_sort.exe -r 10 -n 1000 -c 10 -d 2
# Sorting array of size: 1000 | 10 elements changed in propagation| Domains spawned: 2
## Initial computation
Name                            Avg                Median             Runs     Max time           Min time
static-seq-stdlib-sort          0.1614 ms          0.1440 ms          10       0.3008 ms          0.1199 ms
static-seq-custom-msort         0.1684 ms          0.1659 ms          10       0.1881 ms          0.1549 ms
incr-seq-msort-initial-cons     0.3771 ms          0.3525 ms          10       0.5860 ms          0.3330 ms
incr-par-msort-initial-cons     0.4416 ms          0.4305 ms          10       0.6761 ms          0.3111 ms
js-incr-msort-initial-cons      0.7382 ms          0.6594 ms          10       1.3220 ms          0.6020 ms
current-incr-msort-initial-cons 0.7541 ms          0.6079 ms          10       1.1808 ms          0.5049 ms
## Change propagation
Name                           Avg                Median             Runs     Max time           Min time
incr-par-msort-change-prop     0.06799 ms         0.06544 ms         10       0.09298 ms         0.05412 ms
incr-seq-msort-change-prop     0.09715 ms         0.09906 ms         10       0.1091 ms          0.08511 ms
js-incr-msort-change-prop      0.09951 ms         0.09906 ms         10       0.1149 ms          0.07891 ms
current-incr-msort-change-prop 0.1387 ms          0.1351 ms          10       0.2191 ms          0.1029 ms
## Append propagation
Name                           Avg                Median             Runs     Max time           Min time
incr-seq-msort-append-prop     0.04229 ms         0.04148 ms         10       0.05197 ms         0.03790 ms
js-incr-msort-append-prop      0.04229 ms         0.04100 ms         10       0.05006 ms         0.03600 ms
current-incr-msort-append-prop 0.04439 ms         0.04506 ms         10       0.04792 ms         0.03981 ms
incr-par-msort-append-prop     0.04725 ms         0.04649 ms         10       0.05793 ms         0.03814 ms
+ dune exec --release -- test/merge_sort.exe -r 10 -n 10000 -c 100 -d 2
# Sorting array of size: 10000 | 100 elements changed in propagation| Domains spawned: 2
## Initial computation
Name                            Avg                Median             Runs     Max time           Min time
static-seq-stdlib-sort          1.6531 ms          1.6280 ms          10       1.7671 ms          1.5850 ms
…