@polytypic 2023-07-18
Hmm... I'd recommend just having the benchmark results in this PR rather than in a separate gist.