@dipeshkaphle 2023-07-19
> Hmm... I'd recommend just having the benchmark results in this PR rather than in a separate gist.

I've updated it. I don't know why but when I was trying to add it initially, github told it passed 65000 char limit, so I put it in a gist. But when I edit it and add it, it seems to work.